Refactor : 메인페이지 리팩토링
- Main.tsx : 페이지에는 UI를 불러오는 코드만 사용하도록 처리
- AllPost.tsx : 모든 포스트를 불러오는(무한스크롤)로직이 담긴 UI
- PopularPost.tsx : 인기 포스트를 불러오는(캐러셀)로직이 담긴 UI

그 외, 일부 변수 수정
